split up the implementation into modules
